backend/requirements.txt
fastapi==0.110.0
strawberry-graphql==0.216.0
uvicorn==0.27.1
transformers==4.40.1
torch>=2.0.0
